requirements.txt
fastapi==0.63.0
pycryptodome==3.10.1
pydantic==1.8.1
ecdsa==0.16.1
uvicorn